organize auth folder
