+++ lib/src/app.dart
@@ -1,5 +1,5 @@
 import 'package:book_adapter/src/common_widgets/init_widget.dart';
-import 'package:book_adapter/src/features/auth/auth_checker.dart';
+import 'package:book_adapter/src/features/auth/presentation/auth_checker.dart';
 import 'package:book_adapter/src/features/in_app_update/util/toast_utils.dart';
 import 'package:book_adapter/src/features/library/library_view.dart';
 import 'package:book_adapter/src/features/library/series_view.dart';
