+++ lib/src/features/auth/controller/register_view_controller.dart
@@ -1,8 +1,8 @@
 import 'package:book_adapter/src/controller/firebase_controller.dart';
 import 'package:book_adapter/src/data/failure.dart';
+import 'package:book_adapter/src/features/auth/data/register_view_data.dart';
 import 'package:dartz/dartz.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final registerViewController =
@@ -89,47 +89,4 @@ class RegisterViewController extends StateNotifier<RegisterViewData> {
   }
 }
 
-@immutable
-class RegisterViewData {
-  final String username;
-  final String email;
-  final String password;
-  final String verifyPassword;
-  // null if there is no photo chosen
-  final String? photoUrl;
-  final bool isLoading;
-  final bool isButtonEnabled;
 
-//<editor-fold desc="Data Methods">
-
-  const RegisterViewData({
-    this.username = '',
-    this.email = '',
-    this.password = '',
-    this.verifyPassword = '',
-    this.photoUrl,
-    this.isLoading = false,
-    this.isButtonEnabled = false,
-  });
-
-  RegisterViewData copyWith({
-    String? username,
-    String? email,
-    String? password,
-    String? verifyPassword,
-    String? photoUrl,
-    bool? isLoading,
-    bool? isButtonEnabled,
-  }) {
-    return RegisterViewData(
-      username: username ?? this.username,
-      email: email ?? this.email,
-      password: password ?? this.password,
-      verifyPassword: verifyPassword ?? this.verifyPassword,
-      photoUrl: photoUrl ?? this.photoUrl,
-      isLoading: isLoading ?? this.isLoading,
-      isButtonEnabled: isButtonEnabled ?? this.isButtonEnabled,
-    );
-  }
-//</editor-fold>
-}
