+++ lib/src/features/auth/presentation/reset_password_view.dart
@@ -1,6 +1,6 @@
 import 'package:book_adapter/src/constants/constants.dart';
 import 'package:book_adapter/src/data/failure.dart';
-import 'package:book_adapter/src/features/auth/reset_password_view_controller.dart';
+import 'package:book_adapter/src/features/auth/controller/reset_password_view_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:logger/logger.dart';
