+++ lib/src/features/library/library_view_controller.dart
@@ -1,7 +1,7 @@
 import 'package:book_adapter/src/controller/firebase_controller.dart';
 import 'package:book_adapter/src/controller/storage_controller.dart';
 import 'package:book_adapter/src/data/failure.dart';
-import 'package:book_adapter/src/data/user_data.dart';
+import 'package:book_adapter/src/features/auth/data/user_data.dart';
 import 'package:book_adapter/src/exceptions/app_exception.dart';
 import 'package:book_adapter/src/features/library/data/book_collection.dart';
 import 'package:book_adapter/src/features/library/data/book_item.dart';
