+++ lib/src/features/profile/change_password_view.dart
@@ -1,6 +1,6 @@
 //Handles password change action
 
-import 'package:book_adapter/src/features/auth/reset_password_view.dart';
+import 'package:book_adapter/src/features/auth/presentation/reset_password_view.dart';
 import 'package:flutter/material.dart';
 
 class ChangePasswordView extends StatelessWidget {
